\name{intgrat}
\alias{intgrat}

\title{
Numerical integration using a simple Trapezoidal rule
}
\description{
  Trapezoidal rule is a technique for approximating the definite integral.
}
\usage{
intgrat(f, delta)
}

\arguments{
  \item{f}{		Function values to be integrated. }
  \item{delta}{		Spacing size. }
}

\value{
  \code{intgrat} returns the value of the intergral.
}

\keyword{integration}
