\name{intsim}
\alias{intsim}

\title{Numerical integration using Simpson's rule}
\description{
  Simpson's rule is a method for numerical integration.
}
\usage{
  intsim(f, delta)
}
\arguments{
  \item{f}{		Function values to be integrated. }
  \item{delta}{		Spacing size. }
}

\value{  
  \code{intsim} returns the value of the intergral.
}

\keyword{integration}

