\name{plot.blm}
\alias{plot.blm}
\title{ Plot a blm object}
\description{
Plots the posterior samples for Bayesian linear models
}
\usage{
\method{plot}{blm}(x, \dots)
}
\arguments{
  \item{x}{ a \code{blm} object}
  \item{\dots}{ other options to pass to the plotting functions}
}

\value{Returns a plot.}


\seealso{
\code{\link{blq}}, \code{\link{blr}}
}

\examples{
## See examples for blq and blr
}

