\name{plot.bsad}
\alias{plot.bsad}
\title{ Plot a bsad object}
\description{
Plots the posterior samples for Bayesian semiparametric density estimation using a logistic Gaussian process.
}
\usage{
\method{plot}{bsad}(x, \dots)
}
\arguments{
  \item{x}{ a \code{bsad} object}
  \item{\dots}{ other options to pass to the plotting functions}
}

\value{Returns a plot.}

\seealso{
\code{\link{bsad}}
}

\examples{
## See examples for bsad
}

