\name{plot.bsam}
\alias{plot.bsam}
\title{ Plot a bsam object}
\description{
Plots the posterior samples for Bayesian spectral analysis models.
}
\usage{
\method{plot}{bsam}(x, \dots)
}
\arguments{
  \item{x}{ a \code{bsam} object}
  \item{\dots}{ other options to pass to the plotting functions}
}

\value{Returns a plot.}

\seealso{
\code{\link{bsaq}}, \code{\link{bsaqdpm}}, \code{\link{bsar}}, \code{\link{bsardpm}}
}

\examples{
## See examples for bsaq, bsaqdpm, bsar, and bsardpm
}

