\name{plot.bsamdpm}
\alias{plot.bsamdpm}
\title{ Plot a bsamdpm object}
\description{
Plots the posterior samples for Bayesian spectral analysis models with Dirichlet process mixture error.
}
\usage{
\method{plot}{bsamdpm}(x, \dots)
}
\arguments{
  \item{x}{ a \code{bsamdpm} object}
  \item{\dots}{ other options to pass to the plotting functions}
}

\value{Returns a plot.}

\seealso{
\code{\link{bsaqdpm}}, \code{\link{bsardpm}}
}

\examples{
## See examples for bsaqdpm and bsardpm
}

