\name{plot.fitted.bsad}
\alias{plot.fitted.bsad}
\title{ Plot a fitted.bsad object}
\description{
Plots the predictive density for Bayesian density estimation model using logistic Gaussian process
}
\usage{
\method{plot}{fitted.bsad}(x, \dots)
}
\arguments{
  \item{x}{ a \code{fitted.bsad} object}
  \item{\dots}{ other options to pass to the plotting functions}
}

\value{Returns a plot.}

\seealso{
\code{\link{bsad}}, \code{\link{fitted.bsad}}
}

\examples{
## See example for bsad
}
